\name{Tsdb-package}
\alias{Tsdb-package}
\alias{Tsdb}
\docType{package}
\title{
A collection of utilities for interacting with tsdb.
}
\details{
\tabular{ll}{
Package: \tab Tsdb\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2011-04-16\cr
License: \tab CEG\cr
}
}
\author{
Adrian Dragulescu

Maintainer: Adrian Dragulescu <adrian.dragulescu@constellation.com>
}

\description{
  A collection of utility functions for interacting with tsdb.
}
\keyword{ package }

\examples{

# Check out the individual functions!



}
